renice dish table and its show.
