+++ code/src/components/canteen/CanteenDelete.js
